sched/core: Skip rq lock in try_to_wake_up() when WALT is disabled
When WALT is disabled, do_pl_notif() always returns false, in which case
this bit of code serves no purpose. As a result, this WALT-specific code
spins on acquiring the rq lock in a hot path, wasting CPU time. Compile
it out when WALT is disabled to eliminate the unnecessary overhead.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: Iebc4748e78c6bfd28f989cc052bedcfc87e31412
Signed-off-by: angelomds42 <angelomds42@gmail.com>