UPSTREAM: zstd: fix g_debuglevel export warning
The g_debuglevel variable in debug.c is only used when DEBUGLEVEL is
defined to be above 2. This means by default there's no actual definition
of this in the headers, so sparse is giving the following warning:

lib/zstd/common/debug.c:24:5: warning: symbol 'g_debuglevel' was not declared. Should it be static?

We can use the same check as in the header to remove this if it isn't
going to be used, silencing the warning and removing a small bit of unused
data.

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>
Signed-off-by: Nick Terrell <terrelln@fb.com>
Signed-off-by: mrsrimar22 <mrsrimar22@gmail.com>
Signed-off-by: angelomds42 <angelomds42@gmail.com>