kernel: sched: Rewrite compute_energy() to memoize redundant calculations
Calculating the predicted energy for each CPU candidate is expensive in
large part due to a number of redundant calculations resulting in a total
runtime complexity of O(n**2).

By memoizing the redundant energy computations, the runtime complexity of
calculating energy for all CPU candidates is reduced to O(2n), a
significant performance improvement.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[EmanuelCN: Adapt to msm-4.19 with sultan's help]
Signed-off-by: angelomds42 <angelomds42@gmail.com>